+++ src/common.ts
@@ -1,0 +1,2 @@
+export const noOp = () => {};
+export const NO_ERRORS = Object.freeze([] as string[]);