+++ CHANGELOG.md
@@ -2,6 +2,11 @@ Formula uses [semantic versioning](https://semver.org/).
 
 Formula is currently pre-v1. It's relatively stable, but I expect there will be some breaking changes before v1.
 
+# Unreleased 
+
+- feat: hooks which accept fields now accept null-ish fields
+- fix: useFieldData overload type issue
+
 # 0.0.9
 
 - feat: useFieldData accepts null-ish